+++ src/ClientContextInterface.php
@@ -295,6 +295,7 @@ use Predis\Command\Redis\VADD;
  * @method $this xrange(string $key, string $start, string $end, ?int $count = null)
  * @method $this xread(int $count = null, int $block = null, array $streams = null, string ...$id)
  * @method $this xreadgroup(string $group, string $consumer, ?int $count = null, ?int $blockMs = null, bool $noAck = false, string ...$keyOrId)
+ * @method $this xsetid(string $key, string $lastId, ?int $entriesAdded = null, ?string $maxDeleteId = null)
  * @method $this xtrim(string $key, array|string $strategy, string $threshold, array $options = null)
  * @method $this zadd($key, array $membersAndScoresDictionary)
  * @method $this zcard($key)